+++ parley/src/layout/alignment.rs
@@ -62,10 +62,7 @@ pub(crate) fn align<B: Brush>(
     alignment: Alignment,
     options: AlignmentOptions,
 ) {
-    #[cfg(feature = "accesskit")]
-    {
-        layout.alignment = Some(alignment);
-    }
+    layout.alignment = Some(alignment);
     layout.is_aligned_justified = alignment == Alignment::Justify;
 
     align_impl::<_, false>(layout, alignment, options);